Update ReverseArray.java
The idea is to use recursion and define a recursive function that takes a range of array elements as input and reverses it. Inside the recursive function,

Swap the first and last element.
Recursively call the function with the remaining subarray.

Time Complexity: O(n), the recurrence relation will be T(n) = T(n - 2) + O(1), which can be simplified to O(n).
Auxiliary Space: O(n), as we are using recursion stack.