Arrays
In order to maximize the profit, we need to minimize the cost price and maximize the selling price. So at every step, we keep track of the minimum buy price of stock encountered so far. For every price, we subtract with the minimum so far and if we get more profit than the current result, we update the result

Time Complexity: O(n), as we are traversing the prices[] array only once.
Auxiliary Space: O(1)